remove actix server
