+++ server/Cargo.toml
@@ -28,21 +28,16 @@ phnxbackend = { path = "../backend" }
 phnxtypes = { path = "../types" }
 phnxprotos = { path = "../protos" }
 
-actix-web = "^4.0"
 serde = "1"
 serde_json = "1.0"
 config = "0.15"
-actix-web-actors = "4.2.0"
 actix = "0.13"
 tokio = "1"
-tokio-util = { workspace = true }
-base64 = "0.22"
 thiserror = { workspace = true }
 tracing = { version = "0.1", features = ["log"] }
 tracing-log = "0.2"
 tracing-subscriber = { workspace = true }
 tracing-bunyan-formatter = "0.3"
-tracing-actix-web = "0.7"
 jsonwebtoken = "9"
 zeroize = "1.8.1"
 tls_codec = { workspace = true }
@@ -51,6 +46,7 @@ tonic = { workspace = true }
 tower-http = { workspace = true }
 tokio-stream = { workspace = true, features = ["net"] }
 tower_governor = "0.7.0"
+anyhow = "1.0.98"
 
 [dev-dependencies]
 actix-rt = "^2.7"
